Checkpoint native numeric contracts and proof inputs (unverified)

+++ csharp/src/Tsonic.CSharp.Node/fs/descriptor/read.cs
@@ -13,7 +13,7 @@ public static partial class fs
     /// <param name="length">The number of bytes to read.</param>
     /// <param name="position">The position in the file to start reading from. If null, reads from current position.</param>
     /// <returns>A Task that resolves to the number of bytes read.</returns>
-    public static Task<int> read(int fd, byte[] buffer, int offset, int length, int? position)
+    public static Task<int> read(int fd, byte[] buffer, int offset, int length, long? position)
     {
         return Task.Run(() => readSync(fd, buffer, offset, length, position));
     }
